Separated inner & aggregate responses for FilesEntity in hca-dcp #408
