+++ explorer/app/apis/azul/hca-dcp/common/aggregatedEntities.ts
@@ -1,0 +1,8 @@
+export interface AggregateProject {
+  estimatedCellCount?: number;
+  projectTitle: string[];
+}
+
+export interface AggregateProjectResponse {
+  projects: AggregateProject[];
+}
