+++ explorer/app/apis/azul/hca-dcp/common/entities.ts
@@ -1,17 +1,16 @@
 /**
  * Model of response returned from /index/files API endpoint.
  */
-export interface FilesResponse {
-  files: {
-    contentDescription: string[];
-    format: string;
-    name: string;
-    size: number;
-    url: string;
-    uuid: string;
-  }[];
-  projects: {
-    estimatedCellCount?: number;
-    projectTitle: string[];
-  }[];
+
+export interface FilesEntity {
+  contentDescription: string[];
+  format: string;
+  name: string;
+  size: number;
+  url: string;
+  uuid: string;
+}
+
+export interface FilesEntityResponse {
+  files: FilesEntity[];
 }
