+++ explorer/app/apis/azul/hca-dcp/common/responses.ts
@@ -1,0 +1,7 @@
+import { AzulHit } from "../../common/entities";
+import { FilesEntityResponse } from "./entities";
+import { AggregateProjectResponse } from "./aggregatedEntities";
+
+export type FilesResponse = AzulHit &
+  FilesEntityResponse &
+  AggregateProjectResponse;
